Added FilteredTableWrapper to separate the logic from the actual table
