+++ src/routes/Incidents.js
@@ -1,4 +1,4 @@
-import {Table} from './incidents/table/Table';
+import {FilteredTable} from './incidents/filteredTable/FilteredTable';
 import {Detail} from './incidents/detail/Detail';
 import {useSelector} from 'react-redux';
 import {useGetResource} from '../services/Resources';
@@ -7,7 +7,7 @@ import React from 'react';
 function Incidents() {
   const resources = useSelector((state) => state.auth.resources);
   const incidents = useGetResource(() => resources.incidents());
-  return <Table data={incidents} />;
+  return <FilteredTable data={incidents} />;
 }
 
 export const IncidentDetail = ({match}) => {
