+++ src/routes/incidents/filteredTable/table/filter/categoryFilter/data.js
